requirements-training.txt
descartes==1.1.0
ffio==0.1.0
h5py==3.7.0
imageio==2.22.4
matplotlib==3.6.2
numpy==1.23.4
onnxruntime-gpu==1.13.1
opencv-python==4.6.0.66
Pillow==9.3.0
PyOpenGL==3.1.0
pyrender==0.1.45
python-dotenv==0.21.0
scipy==1.9.3
Shapely==1.8.5.post1
torch==1.13.0
torchvision==0.14.0
wandb==0.13.5
tqdm==4.64.1
pandas==1.5.1